Task Lister
It is a task lister (To-do list) that enables a user to:
As a user, I should be able to type a task into the input field.
As a user, I should be able to click some form of a submit button.
As a user, I expect to see the task string that I provided appear in the DOM after the submit button has been activated.
A delete function that will remove tasks from your list
Ability to edit tasks.